+++ pkg/corerp/renderers/container/manifest.go
@@ -33,6 +33,7 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/util/strategicpatch"
 )
 
@@ -227,4 +228,35 @@ func patchPodSpec(sourceSpec *corev1.PodSpec, patchSpec []byte) (*corev1.PodSpec
 	}
 
 	return patched, nil
+}
+
+func mergeLabelSelector(base *metav1.LabelSelector, cur *metav1.LabelSelector) *metav1.LabelSelector {
+	if base == nil {
+		base = &metav1.LabelSelector{}
+	}
+
+	return &metav1.LabelSelector{
+		MatchLabels:      labels.Merge(base.MatchLabels, cur.MatchLabels),
+		MatchExpressions: append(base.MatchExpressions, cur.MatchExpressions...),
+	}
+}
+
+func mergeObjectMeta(base metav1.ObjectMeta, cur metav1.ObjectMeta) metav1.ObjectMeta {
+	return metav1.ObjectMeta{
+		Name:        cur.Name,
+		Namespace:   cur.Namespace,
+		Labels:      labels.Merge(base.Labels, cur.Labels),
+		Annotations: labels.Merge(base.Annotations, cur.Annotations),
+	}
+}
+
+func getObjectMeta(base metav1.ObjectMeta, appName, resourceName, resourceType string, options renderers.RenderOptions) metav1.ObjectMeta {
+	cur := metav1.ObjectMeta{
+		Name:        kubernetes.NormalizeResourceName(resourceName),
+		Namespace:   options.Environment.Namespace,
+		Labels:      renderers.GetLabels(options, appName, resourceName, resourceType),
+		Annotations: renderers.GetAnnotations(options),
+	}
+
+	return mergeObjectMeta(base, cur)
 }
